Add bar button segue to new view controller
